src/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-simple-history==3.3.0
django-mdeditor==0.1.20
Markdown==3.4.3
mysqlclient==2.1.1
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
gunicorn==20.1.0
